+++ src/command_modules/azure-cli-acr/HISTORY.rst
@@ -3,6 +3,11 @@
 Release History
 ===============
 
+2.1.1
++++++
+* Add --with-secure-properties flag in 'acr build-task show' command.
+* Add 'acr build-task update-build' command.
+
 2.1.0
 +++++
 * BREAKING CHANGE: Update '--no-push' to a pure flag in 'acr build' command.